moved PostsContent to utils. may not be the final resting place
